[FE/BE][:hammer_and_wrench: FEAT]: 일기 삭제 시 댓글 남는 현상 수정 + AOS 애니메이션 일부 적용 + 그래프 캘린더 이모지 해당 달 외 표시 제거 + 댓글 삭제 api 추가
