+++ src/main/frontend/package.json
@@ -13,6 +13,7 @@
     "@types/react-dom": "latest",
     "@types/react-router-dom": "^5.3.3",
     "acorn": "^8.8.0",
+    "aos": "^2.3.4",
     "axios": "^0.27.2",
     "canvas": "^2.9.3",
     "chart.js": "^3.8.2",
@@ -65,6 +66,7 @@
     "@jest/globals": "^28.1.3",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
+    "@types/aos": "^3.0.4",
     "@types/jest": "^28.1.6",
     "@types/react-helmet": "^6.1.5",
     "@types/sentiment": "^5.0.1",
