+++ src/main/frontend/src/components/molecules/IconSet.tsx
@@ -8,10 +8,18 @@ export interface IconSetProps {
   content: string;
 }
 
-function IconSet({ index, icon, title, content }: IconSetProps) {
+interface IconOrderSetProps extends IconSetProps {
+  order: number;
+}
+
+function IconSet({ index, icon, title, content, order }: IconOrderSetProps) {
   return (
-    <Stack sx={{ alignItems: "center", width: 300 }} spacing={5}>
-      {index && (
+    <Stack
+      sx={{ alignItems: "center", width: 300 }}
+      spacing={5}
+      data-aos='fade-down'
+      data-aos-delay={100 * (order + 1)}>
+      {index >= 0 && (
         <Typography
           sx={{
             fontWeight: 700,
