+++ src/main/java/com/narang/web/service/CommentService.java
@@ -9,36 +9,41 @@ import java.util.List;
 
 @Service
 public class CommentService {
-    private CommentRepository commentRepository;
+  private CommentRepository commentRepository;
 
-    @Autowired
-    CommentService(CommentRepository commentRepository) {
-        this.commentRepository = commentRepository;
-    }
+  @Autowired
+  CommentService(CommentRepository commentRepository) {
+    this.commentRepository = commentRepository;
+  }
 
-    public List<Comment> findAll() {
-        return commentRepository.findAll();
-    }
+  public List<Comment> findAll() {
+    return commentRepository.findAll();
+  }
 
-    public List<Comment> findByDid(String did) {
-        return commentRepository.findByDid(did);
-    }
+  public List<Comment> findByDid(String did) {
+    return commentRepository.findByDid(did);
+  }
 
-    public Comment findById(String id) {
-        return commentRepository.findById(id).orElseThrow();
-    }
+  public Comment findById(String id) {
+    return commentRepository.findById(id).orElseThrow();
+  }
 
-    public String insert(Comment comment) {
-        Comment newComment = commentRepository.insert(comment);
-        return newComment.getId();
-    }
+  public String insert(Comment comment) {
+    Comment newComment = commentRepository.insert(comment);
+    return newComment.getId();
+  }
 
-    public Comment update(Comment comment) {
-        return commentRepository.update(comment);
-    }
+  public Comment update(Comment comment) {
+    return commentRepository.update(comment);
+  }
 
-    public Boolean deleteById(String id) {
-        commentRepository.deleteById(id);
-        return true;
-    }
+  public Boolean deleteById(String id) {
+    commentRepository.deleteById(id);
+    return true;
+  }
+
+  public Boolean deleteByDid(String did) {
+    commentRepository.deleteByDid(did);
+    return true;
+  }
 }
